third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/empty-filters-member.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
    return assert_promise_rejects_with_message(
        requestDeviceWithKeyDown({filters: []}),
        new TypeError('Failed to execute \'requestDevice\' on \'Bluetooth\': ' +
            '\'filters\' member must be non-empty to find any devices.'));
}, 'An empty |filters| member should result in a TypeError');
</script>
